ForeAITgBot/pyproject.toml
[project]
name = "ForeAITgBot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytelegrambotapi>=4.26.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
